refactor: modularize the typing for qchem results
